functions/thumbnail_generator/requirements.txt
functions-framework==3.*
google-cloud-storage==2.14.0
google-cloud-firestore==2.14.0
Pillow==10.2.0
cloudevents==1.10.1

